requirements.txt
torch
torchvision
numpy
pillow
diffusers
transformers
accelerate
xformers